Add classes to support the Plugin API
Adds the CronOption and CronOptions classes. An instance of the
CronOptions class is (currently) passed as the `subject` arg for the
`onCronOptionsList` event. The plugin subscriber is supposed to use the
addOptions method of CronOptions to pass supported jobs.
The changes in the SelectModel and a test plugin implementation follow
this commit.