Fix asset tracking in CronjobTable
- Adds the `_getAssetName()` method to override the parent method which
defaulted to an asset name of #__cronjobs.${jobId}. This resulted in an
unwanted extra entry in the #__assets table that was never referenced.
Instead, a com_cronjobs.cronjob.${jobId} entry was created when updating
asset permissions and referenced by com_cronjobs when checking for
permissions. This entry should be created at cronjob creation now.